Refactored the display chunk generation to facilitate both hot and cold paths at the same time
